Extract sort function and make generic
